Extracted agent file management to its own service
